+++ src/state.c
@@ -301,7 +301,7 @@ static void ss_register_func(ss_func **root, void (*func)(void))
 	{
 		if (next->func == func && next->tag == ss_current_tag)
 		{
-			logerror("Duplicate save state function (%d, 0x%x)\n", ss_current_tag, (int)func);
+			logerror("Duplicate save state function (%d, 0x%x)\n", ss_current_tag, (size_t)func);
 			exit(1);
 		}
 		next = next->next;
